SSR fixes based on Coderabbit
